[fix] Improve upload resilience coverage
Cover FileSlicer cancellation cleanup paths and address the async disposal warning reported by SonarCloud.

Made-with: Cursor

+++ src/ByteSync.Client/Services/Communications/Transfers/Uploading/UploadAttemptTimeoutPolicy.cs
@@ -19,7 +19,7 @@ public static class UploadAttemptTimeoutPolicy
         }
         
         var staleChunkPenalty = ComputeStaleChunkPenaltySeconds(sliceLengthBytes, currentChunkSizeBytes);
-        timeoutSec += (attempt - 1) * RetryGrowthSeconds + staleChunkPenalty;
+        timeoutSec += (long)(attempt - 1) * RetryGrowthSeconds + staleChunkPenalty;
         
         return (int)Math.Clamp(timeoutSec, AttemptTimeoutFloorSeconds, AttemptTimeoutCeilingSeconds);
     }
